Update patch set 1
Patch Set 1:

(appending to the previous comment) I did just bump the upper constraint to fix the test because I knew you all didn't like bumping the requirement (and justified that to myself because train is old, ovn 2.11 is old and even with train-based releases of RHEL OSP we use 2.13, and the failures we see are on py27 which is also dead-ish). But really, when we commit the other race condition fixes in networking-ovn, it really will make this fix a requirement as in at least one particular case, we'd go from failing occasionally to failing consistently.

Patch-set: 1