Update patch set 1
Patch Set 1:

> do you mind answer the qestions 

not a problem but on second thought: s3transfer is a dependency of boto3, and boto3 is already there and runs fine. If we are not adding any explicit dependency onto s3transfer, the requirement check of the CI pipeline will just run fine.

What do you prefer?
- option 1 where we go ahead with this change with the right commit message ?
- option 2 where we just discard this change, s3transfer will be installed "silently" as it is  done today for components that depend on boto3

Patch-set: 1