Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> Why is python-tripleoclient is in tripleo-validations requirements/test-requirements? I see it was added recently in https://github.com/openstack/tripleo-validations/commit/59fdd2d1ee44403563ec9dcab2a17038d446dc0f. I've no idea why.

This is a good point. At the time we were still not certain how is the CLI going to work out. Note the osc-lib inclusion in the same patch. The python-tripleoclient was included for the same reason. Now that there is no need for it, I believe it's safe to get rid off.

Thank you.

Patch-set: 2