Update patch set 10
Patch Set 10:

I've tried to reproduce the unit test slowdown with various sqlalchemy 1.4 versions. But I cannot see the slowdown locally. Moreover a the nova unit test are a bit faster with 1.4 than with 1.13.23. I've also checked the amount of log the job generates but that also did not changed significantly between 1.3 and 1.4. So at the moment I don't know why the unit test times out in this patch frequently but runs well with 1.3 on the gate and with 1.4 in local for me.

Patch-set: 10
Reviewer: Gerrit User 9708 <9708@4a232e18-c5a9-48ee-94c0-e04e7cca6543>