Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-2
> 
> hi,
> 
> this list was originally alphabetized, but was intentionally disorganized so that multiple reviews for things like the oslo projects would not conflict with eachother as often in git (preventing the need to rebase).
> 
> https://github.com/openstack/requirements/commit/26b6c243fc8039483537810f6e9cc20a3c56055c

I find hard to understand the logic behind, cryptic would be an understatement. Still, this makes me ask the following question: if that file is machine generated, why is not starting with a comment mentioning that and the command someone needs to run in order to rebuild it? Like:

```
# Generated via: ...
```

I should remark that I read https://docs.openstack.org/project-team-guide/dependency-management.html and the Adding new dependency and that section is highly outdated mentioning py27 which was dropped long time ago. Once things gets clear I would be happy to update that doc part, especially as I want to add few more test dependencies.

Patch-set: 2