Revert "Add neutron to stable upper-constraints"
This reverts commit 1c51379ed4dd3bbe97601c4b8eae9d058b134cb0.

Reason for revert: Neutron is being used also by other projects (like deployment ones) that need to comply with u-c and at the same time need to install specific service versions (with SHAs).
New pip resolver does not allow to have specific SHAs in constraints and thus it needs to be put into requirements, which results in conflict and being unable to install neutron and use u-c at the same time.

Change-Id: I16b30dcce2db84351460cc2c362a42a0b8532b79