Update patch set 1
Patch Set 1:

Not sure if we really can do this now. There are projects that use neutron as a library. For stable branches, we had run into issues in the past where some of these stable branches started breaking because they were unknowingly pulling in a neutron a few releases newer than what they had shipped with. The fix for that was to start adding neutron to stable constraints.

Patch-set: 1
CC: Gerrit User 11904 <11904@4a232e18-c5a9-48ee-94c0-e04e7cca6543>