Add autopage to global-requirements
autopage is a library for automatically invoking 'less' when needed in
command-line programs and piping the output through it. It is wanted to
improve the user experience of python-openstackclient and other
cliff-based command line tools.

The library is actively maintained (by me), and licensed under the
Apache 2.0 license. Distro packages for Fedora/CentOS and Ubuntu are
available in Copr and PPA, respectively:

https://copr.fedorainfracloud.org/coprs/zaneb/autopage/packages/
https://launchpad.net/~zaneb/+archive/ubuntu/autopage

Change-Id: Ic773519e5ab3e940cb23a96b99295f7eb8986691