Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Is there a reason this needs to be a library, there seems to be a bunch of things you can do natively in python for datetime ranges?  https://stackoverflow.com/questions/993358/creating-a-range-of-dates-in-python  I will say the library does seem cleaner, just don't want to turn this into nodejs :P

I read the same stackoverflow and other places. Also the same question that a friend posed. Well, if we write the code, it is more to maintain, document, and to guarantee that it works (via unit testings normally). By adopting a library, I know that there is a solid solution, tested, and used out there. Moreover, it is less code to maintain throughout time.

I personally see no reason to replicate others work. So, if there is already a code (library) that does exactly what I need, it is well maintained, the code is neat, documented, I do not see the rationality not to adopt it.

If people are opposed to adopting it, I would rewrite it, or part of the solution in CloudKitty. However, I do not see that as the best approach, specially in the opensource world.

Patch-set: 5