Update patch set 2
Patch Set 2:

> would be nice to know what happens if you don't want to maintain autopage anymore).

Fair question, it's a bit of a chicken-and-egg problem. If it is in use somewhere then maybe someone will be interested, but somebody has to go first.

I don't expect keeping it working with new Python releases (as opposed to adding new features) to be very much work, if any, and I have (work) time available to do it, so I am comfortable accepting the responsibility of being a sole maintainer for now.

Patch-set: 2