bumping validations-libs version
Validations libs recently underwent several important changes
that are now included in the 1.2.0 release.

The upper constraints have to be updated to ensure that proper
version of the package is pulled for tests and other purposes.

The backport of patch in python-tripleoclient[0] depends on code in 
the 1.2.0 release of validations libs.

But because of the upper constraints the tox-* jobs would consistently fail,
as the 1.1.0 doesn't include the relevant patches for the validations-libs CLI.
The original bug that led to this patch,
was found in 16.2 and the primary fixes were targeting it.[2]

Furthermore,some featueres, planned for wallaby, are only in the validations-libs>=1.2.0,
like the metadata expansion.

[0]https://review.opendev.org/c/openstack/python-tripleoclient/+/794577
[1]https://review.opendev.org/c/openstack/validations-libs/+/795093
[2]https://review.opendev.org/c/openstack/validations-libs/+/789770

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: Ie3f0be9c39e030c2600f0723467d244699fbf173
(cherry picked from commit fcfb03f51322bb48cc1cad02a6f5b0f37926ec7a)