Update patch set 1
Patch Set 1: Workflow-1

> Patch Set 1: Code-Review-1
> 
> I think I need more details as to why we need to update a lib on a stable branch before we can do so.
> 
> https://docs.openstack.org/project-team-guide/dependency-management.html#upper-constraints

The backport of patch in python-tripleoclient[0] depends on code in 1.2.0 release of validations libs.
But because of the upper constraints the tox-* jobs would consistently fail, as the 1.1.0 doesn't include the relevant patches for the validations-libs CLI.

Now the patch itself isn't substantial, but without it the newly exposed validations-log-path arg won't behave properly. The original bug that led to this patch, was found in 16.2 and the primary fixes were targeting it, 

If we decide there is no need for this functionality[2] in Wallaby, or earlier, there will be no need to change the constraints. But as it stands now, I don't have a firm guidance on that from the rest of our team. 

So I guess until then this will have to wait. For the better probably, wouldn't want to break things. 

[0]https://review.opendev.org/c/openstack/python-tripleoclient/+/794577
[1]https://review.opendev.org/c/openstack/validations-libs/+/795093
[2]https://review.opendev.org/c/openstack/validations-libs/+/789770/23

Patch-set: 1
Reviewer: Gerrit User 32926 <32926@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=-1