Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> I think I need more details as to why we need to update a lib on a stable branch before we can do so.
> 
> https://docs.openstack.org/project-team-guide/dependency-management.html#upper-constraints

It appears that the functionality is targeting wallaby after all, so we will need to updated the constraints accordingly. 
Furthermore, there other items, included in the validations-libs>=1.2.0 like the metadata expansion[3], that are intended for the 17, so we need this for them as well.

Patch-set: 1
Reviewer: Gerrit User 32926 <32926@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=0