Remove libvirt-python from upper-constraints
The python bindings for libvirt are strictly related to the libvirt
system packages installed in each distribution.
When installing libvirt-python from pip and forcing an installation
of a version higher from the one already installed by the system
will cause the error:
"ERROR: Cannot uninstall 'libvirt-python'. It is a distutils installed
project and thus we cannot accurately determine which files belong to
it which would lead to only a partial uninstall."
To avoid this issue we remove libvirt-python from upper-constraints
leaving entirely to the projects how to handle its version directly
from their requirements.

Change-Id: Ic6174bfe7b4c911664631df77cf9b49a895c284d