Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> expected, but only nova is good

I'm looking at this and have found that we're hitting this issue again:

https://github.com/harlowja/fasteners/issues/36

due to this commit which was released in fasteners==0.15:

https://github.com/harlowja/fasteners/commit/467ed75ee1e9465ebff8b5edf452770befb93913

Inside of the nova test fixture, we are somehow getting the unpatched version of threading.current_thread instead of the eventlet version that we need. This implies that at the time of the creation of our ReaderWriterLock inside the fixture, the threading module is not eventlet monkey patched:

https://github.com/harlowja/fasteners/issues/36#issuecomment-760402680

I don't know how or why this is the case, same as bnemec mentioned here:

https://github.com/harlowja/fasteners/issues/36#issuecomment-771772897

I'm going to keep looking but will need help from other nova devs who know more about eventlet.

Patch-set: 2
CC: Gerrit User 4690 <4690@4a232e18-c5a9-48ee-94c0-e04e7cca6543>