Drop unbuildable packages from global-requirements
These are no longer buildable on a modern python3 environment, the only
thing that they didn't show up as gate failures so far is that we are
running on the opendev CI with a set of cached wheel builds, where
versions of these packages were built with an older setup some time age
when things still worked. This will no longer be the case when switching
to python3.10. There also seem to be no current users of these packages
anymore. Affected are:

- anyjson
- pystache
- suds-jurko

See [0] for a test build without the wheel cache in use.

[0] https://review.opendev.org/813292

Change-Id: Ic9b4200ad98879c9d2bd4c0544ba9f1f3fb412f9