Constrain 'coverage' for python 3.9
Yoga supports py3.6, 3.8, and 3.9 and testing is supposed to be done
in at least 3.6 and 3.9 [0].  For the 'coverage' package, the
following is currently in upper-constraints.txt:

  coverage===6.2;python_version=='3.6'
  coverage===6.3.1;python_version=='3.8'

(We have to do something like this because python 3.6 is EOL and
packages we consume are dropping support for it.)

This leaves the package unconstrained for python 3.9, and the pattern
is repeated throughout the file for other packages.  It also leaves
python 3.7 unconstrained.  (Technically, OpenStack doesn't support 3.7,
but I'm not sure that should mean "anything goes" in the u-c.)

The point of this patch is to get the discussion going about what
pattern we should be using in the u-c file to handle this situation.

[0] https://governance.openstack.org/tc/reference/runtimes/yoga.html

Change-Id: I4c644e473c0be569f36056e1505bf0eac76df51a