DNM: Pin nonexistent lockfile version
The lockfile package is abandoned, and we removed it from our
requirements in 2015, but apparently we still indirectly depend on
it since it's still in our constraints list. Try forcing a
nonexistent version and seeing if we can get sufficient clues from
the resulting build failure to figure out what's requiring it.

Change-Id: I7b07c35847fa34599a00b28104a0ff77d0942c25