Allow backports to have arbitrary python_version markers
There are a number of backport packages in place for packages in stdlib
that are still evolving. One such package ins 'importlib-metadata' which
is a backport of 'importlib.metadata'. The version of
'importlib.metadata' found in Python 3.10 provides a function,
'packages_distributions', that is not found in older Python versions.
Anyone that wants to use this function on e.g. Python 3.9 needs to use
the backport library (version 4.4, specifically). However, anyone that
does not need this additional functionality can avoid using the backport
except on very old Python versions. It should be possible for libraries
to determine exactly when they stop requiring the backport library. Make
that possible by allowing the 'python_version' marker to differ for
these libraries.

Change-Id: I85501b4bff97d1c1e1873c3329ef998a8e501134
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>