[stable-only] Cap importlib-metadata on Python 3.7
In the recently released importlib-metadata 5.0.0 package a number of
breaking changes where made to the interface (in fact that was the only
change for the release removing a bunch of things tagged as deprecated).
Since Python 3.7 depends on this library to function for querying
package metadata these breaking API changes cause anything using
stevedore to fail when 5.0.0 is installed. Supporting a compatible
interface across multiple versions of importlib is difficult to to the
nature of these breaking api changes. The least risky stable branch
change to make here is to cap the version, especially as this is a 3.7
only requirement this isn't a concern for master or any newer release.
To workaround these issues this commit sets a version cap on
importlib-metadata to avoid using the version which is incompatible
with stevedore.

The stevedore cap is being added in change
I0db3624fcdb50eea4ff039e65456ef6e91c99c48

Change-Id: Id696f42db7f47976aaf2f42bad3ca3fd8202be27