Bump pyOpenSSL and cryptography upper constraint
Since Zed is also deployed on CentOS Stream 9, it would be best to
at least use the same OpenSSL and derivatives releases as master,
especially following the latest changes in this world (CVE fixes and the
like).

This bump will allow to remove a hard pinning of the "cryptography"
library, allowing to make the whole project more secure and up-to-date
regarding OpenSSL ecosystem.

Change-Id: I1562e0fd60a6340e26a77a2ab6102595dc7ee979