Update patch set 1
Patch Set 1:

Reintroducing 4.0.0 into constraints with an environment marker to keep using <4 with Python 3.8 makes sense if the tooz maintainers still intend to drop 3.8 completely this cycle. In the meantime, reverting the constraints bump is a quick temporary workaround while we wait for them to decide how they want to proceed.

Patch-set: 1