Bump oslo.db to 12.3.2
We are currently blocked from upgrading to oslo.db 13.x since that
removes support for things like sqlalchemy-migrate, which unfortunately
are still being used by a few projects. However, there's an annoying bug
that is causing spurious warning messages during test runs in various
projects. This has been fixed on the most recent stable branch [1], so
we can at least bump our upper constraint to drag this in while waiting
for a chance to use 13.x.

Change-Id: I37fe5174dbd7deadb350aa9956581a944202ffbc
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>