[DNM] Testing multinode multiversion python constraints generation
The basic theory here is that we can deploy a node with known versions
of python installed.
On each node generate a set of constraints for each known version
then on a control node merge and freeze those generated constraints
and ultimately propose that as a new upper-constraints.txt

This is an enhancement as we're using *real* versions of python can
support SLURP runtimes more accurately.  At the other end as can
incorporate unreleased runtimes which will eventually be needed.

This is very work-in-progress.
TODO (in no particular order):
 - Verify pause/return can be used for inter node communications
   - It's possible that this would be better structured with explict
     requires/provides and job dependencies ... time and review will
     tell
 - Actually generate valid venvs and constraints
 - Split the merge function out of generate
 - Incorporate work from https://review.opendev.org/q/topic:requirements-ops
   for better debugging and human readability
 - review/fix/merge into openstack/project-config
 - run in parallel with existing implementation
 - #profit/relax

Change-Id: Ic1501ceaaf701e3b0e294a8535ae2c17d07c1133