Revert Pillow version pin to 9.5.0
The Pillow bump broke the blockdiag library, it is impossible to
render diagrams, and that breaks the execution of the docs job
when using tools like seqdiag or blockdiag.
This patch set version of Pillow back to 9.5.0 until a valid
alternative can be found.
At the moment the best alternative seems to be PlantUML
https://plantuml.com/

For more details see [1]

[1] https://bugs.launchpad.net/ironic/+bug/2026345

Change-Id: I19749cd7f60293424e5823a370a3ef5a0a893a99