Partial Revert "Updated from generate-constraints"
This partially reverts commit c40d5da7fd9ae10ce0c38a548ab876db3f08e95d.

In the above commit we added a version specifier for funclibparser
due to a) there was an 2.0.0a0 release available ; abd b) caution
around potentially breaking docs builds.

This reverts that and part of the change and should be applied early in
the D/2024.2 cycle

Change-Id: I51ed0e4c181b136e8291c373847ba3d12664451a