[DNM] Create a merge-constraints tool
Currently we have a tool call generate-constraints that operates like so:

foreach supplied python_version
  Create a venv
  Install global-requirements in that venv
  generate a pip frezze

foreach freeze combine
  Combine into a single set of constrains.

Use the 'proposal' framework in project-config to generate a gerrit
change wit the new constraints

There is also the ability to simulate a python_version, but this isn't
accurate.  The major limitation is we can only generate accurate
constraints for python versions supplied by the distribution.

This change and the following changes will alter that process.

Step one: (this change)
 Move the generation and combination code into a library
 Create a new tool call merge-constraints that will take individually
   create freeze outputs and combine them to generate a single
   constraints file.

This is done without disrupting the existing generate-constratints tool

Step two:

Add a multinode job that will use ansible to generate a freeze for all
available python versions.  This is multinode so we can add new OSs and
python releases somewhat trivially.

At the time of writing:
  Use ubuntu-focal to generate py3.8 and py3.9 freezees
  Use ubuntu-jammy to generate py3.10 and py3.11 freezees

Step three:

Add another job that will use ansible to take the freezes generated in
step two and combine them into a single constraints file

Step four, and beyond: (to be written)

Add the new jobs into the proposal framework alongside the existing
generate-constraints tool for comparison.  Ensure that the bulk of the
tools and job definitions are in openstack/requirements with only the
proposal logic in openstack/project-config, it needs to be there to
access the zuul secrets.Disable and remove the older tool once we're
happy with the new tool.

Change-Id: If4f9d1adb9c259dc0fbe067ff6cf9e007d697acf