Update patch set 1
Patch Set 1:

We've typically considered that a bug in the consuming project (kuryr). Any directly imported module should have its package listed in the project's requirements. Once upon a time (2014?) we implemented an experimental checker tool to find these, but it was based on calling some internal pip functions and so was ultimately abandoned without the necessary rewrite. It would still be a good thing to find an automated means of checking, as it clearly does impact our projects from time to time.

Patch-set: 1
CC: Gerrit User 5263 <5263@4a232e18-c5a9-48ee-94c0-e04e7cca6543>