Pin bcrypt until passlib is fixed/removed
Currently there is an incompatibility with the latest version of passlib and
the version of bcrypt we have specificed in the upper-constraints [1]. Passlib
also apears to be unmaintained [2] at this point and we're investigating the
removal of the library from Keystone. For the time being this pins bcrypt to a
version that is known to work with the latest passlib that is available.

[1]: https://foss.heptapod.net/python-libs/passlib/-/issues/190
[2]: https://foss.heptapod.net/python-libs/passlib/-/issues/187

Change-Id: Id6330467d95216288fd3722b83a6eb80761c17ab