Group requirements into sections and sort
We add the script used to do this. This script will grow in complexity
in future changes as we add logic to e.g. remove old version constraints
that are no longer necessary. There's probably more sorting to be done
here, but that can always be tacked on with a follow-up.

Psst, reviewers: you can ensure we haven't lost/gained any dependencies
with this little script:

  #!/usr/bin/env python3
  import pathlib
  import subprocess
  import sys

  previous = subprocess.check_output(
      ['git', 'show', 'HEAD~:global-requirements.txt']
  ).decode(sys.stdout.encoding)
  current = pathlib.Path('global-requirements.txt').read_text()

  assert sorted(
      x for x in previous.split('\n') if x and not x.startswith('#')
  ) == sorted(
      x for x in current.split('\n') if x and not x.startswith('#')
  )

We haven't checked _this_ in since it's trivial and not broadly useful
here.

Change-Id: If836960167fe66534034a8a512ea63553649c3ff
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>