Update cross jobs to be explicit about which branch to checkout
As some projects create unmaintained/zed branches and remove the
stable/zed branches it becomes slightly difficult to gate on these
projects.  At the moment there isn't a good, documented, way to do
branch fallback so as a project removes stable/zed we end up cross
gating against master which isn't the intent.

This change switches from the required-projects from a simple list of
strings to list of dictionaries.  Each item contains an explict branch
to checkout for the gated project.  This is correct at the time of
writing but may change.  If a project has both stable/zed and
unmaintained/zed I selected the latter.

Also an unrelated style change for the files list in a couple of jobs so
that the file is self consistent.

Change-Id: I61af5056c3bf25c52712ba559c8e99cffcc123ca